Persist sendgrid events
This uses our data structure (how we persist events to be able to show
them in Gridlook).

TODO: Write more complete tests.

In the future if we want this project to be more generic so it has more
value for others, we should probably make it possible to have a callback
that others can use to be able to persist data in whatever format they
like.