Fix broken test
Since we use the Basic Auth plug in the events controller it will be in
use in the tests and since we didn't set the credentials the test would
break.

So another way of doing this is just to run the plug in
production. But testing basic auth has it's value so I prefer this
solution.